requirements.txt
nats-py
msgspec
